Fix types
